Solve Spark versioning problem
Copy all called internal spark functions into relevant
class objects, which have been moved out of the spark
package into org.broadinstitute.hail.sparkextras.  The
only things left in the Spark package are:
  PartitionedDataFrameReader
  PartitionedParquetRelation

This addresses version problem because the functions are
copied into the Hail codebase and internally consistent.
