+++ src/main/scala/org/apache/spark/sql/ExposedSpark.scala
@@ -1,0 +1,5 @@
+package org.apache.spark.sql
+
+object ExposedSpark {
+  class ExposedDataFrameReader(sqlContext: SQLContext) extends DataFrameReader(sqlContext)
+}
