+++ src/main/scala/org/broadinstitute/hail/driver/ImportAnnotations.scala
@@ -1,10 +1,10 @@
 package org.broadinstitute.hail.driver
 
-import org.apache.spark.rdd.OrderedRDD
 import org.apache.spark.sql.Row
 import org.broadinstitute.hail.Utils._
 import org.broadinstitute.hail.annotations.Annotation
 import org.broadinstitute.hail.expr._
+import org.broadinstitute.hail.sparkextras.OrderedRDD
 import org.broadinstitute.hail.utils.{TextTableOptions, TextTableReader}
 import org.broadinstitute.hail.variant._
 import org.kohsuke.args4j.{Argument, Option => Args4jOption}
