+++ src/main/scala/org/broadinstitute/hail/sparkextras/OrderedLeftJoinRDD.scala
@@ -1,6 +1,7 @@
-package org.apache.spark.rdd
+package org.broadinstitute.hail.sparkextras
 
 import org.apache.spark._
+import org.apache.spark.rdd.RDD
 import org.broadinstitute.hail.Utils._
 import org.broadinstitute.hail.utils.BufferedAdvanceableOrderedPairIterator
